+++ unstructured/ingest/runner/__init__.py
@@ -46,4 +46,5 @@ __all__ = [
     "sharepoint",
     "slack",
     "wikipedia",
+    "writer_map",
 ]