+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## 0.10.12-dev0
+
+### Enhancements
+
+### Features
+
+### Fixes
+
+## 0.10.11
+
+### Enhancements
+
+* Bump unstructured-inference
+  * Combine entire-page OCR output with layout-detected elements, to ensure full coverage of the page (0.5.19)
+
+### Features
+
+* Add in ingest cli s3 writer
+
+### Fixes
+
+* Fix a bug where `xy-cut` sorting attemps to sort elements without valid coordinates; now xy cut sorting only works when **all** elements have valid coordinates
+
 ## 0.10.10
 
 ### Enhancements